+++ cli/cli/defaults/defaults.go
@@ -6,15 +6,10 @@
 package defaults
 
 import (
-	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_cli_version"
 	"github.com/sirupsen/logrus"
 )
 
 const (
-	kurtosisDockerOrg = "kurtosistech"
-
-	DefaultJavascriptReplImage = kurtosisDockerOrg + "/javascript-interactive-repl:" + kurtosis_cli_version.KurtosisCLIVersion
-
 	// If this version is passed to the engine, the engine will use its default version
 	DefaultAPIContainerVersion = ""
 )